Move ExternalizableTypeElement and TypeElementFactory.
