+++ cr-examples/onnx/src/main/java/oracle/code/onnx/ir/OnnxType.java
@@ -25,7 +25,8 @@
 
 package oracle.code.onnx.ir;
 
-import jdk.incubator.code.ExternalizableTypeElement;
+import jdk.incubator.code.dialect.ExternalizableTypeElement;
+import jdk.incubator.code.dialect.TypeElementFactory;
 
 import java.util.ArrayList;
 import java.util.List;
