+++ cr-examples/spirv/src/main/java/intel/code/spirv/SpirvType.java
@@ -25,7 +25,7 @@
 
 package intel.code.spirv;
 
-import jdk.incubator.code.ExternalizableTypeElement;
+import jdk.incubator.code.dialect.ExternalizableTypeElement;
 
 public abstract sealed class SpirvType implements ExternalizableTypeElement permits PointerType, StorageType {
 }
