+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/bytecode/UnresolvedType.java
@@ -25,7 +25,7 @@
 
 package jdk.incubator.code.bytecode;
 
-import jdk.incubator.code.ExternalizableTypeElement;
+import jdk.incubator.code.dialect.ExternalizableTypeElement;
 import jdk.incubator.code.TypeElement;
 import java.util.List;
