+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/dialect/java/ArrayType.java
@@ -30,7 +30,6 @@ import java.lang.invoke.MethodHandles.Lookup;
 import java.lang.reflect.GenericArrayType;
 import java.lang.reflect.Type;
 
-import jdk.incubator.code.ExternalizableTypeElement;
 import jdk.incubator.code.dialect.java.impl.JavaTypeUtils;
 
 /**
