+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/dialect/java/ClassType.java
@@ -25,7 +25,6 @@
 
 package jdk.incubator.code.dialect.java;
 
-import jdk.incubator.code.ExternalizableTypeElement;
 import jdk.incubator.code.dialect.java.impl.JavaTypeUtils;
 
 import java.lang.constant.ClassDesc;
