+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/dialect/java/JavaType.java
@@ -31,7 +31,7 @@ import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodHandles.Lookup;
 import java.lang.reflect.*;
 
-import jdk.incubator.code.ExternalizableTypeElement;
+import jdk.incubator.code.dialect.ExternalizableTypeElement;
 import jdk.incubator.code.dialect.java.WildcardType.BoundKind;
 
 import java.util.List;
