+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/dialect/java/TypeVariableType.java
@@ -25,8 +25,7 @@
 
 package jdk.incubator.code.dialect.java;
 
-import jdk.incubator.code.ExternalizableTypeElement;
-import jdk.incubator.code.TypeElement;
+import jdk.incubator.code.dialect.ExternalizableTypeElement;
 import jdk.incubator.code.dialect.java.impl.JavaTypeUtils;
 
 import java.lang.constant.ClassDesc;
