+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/parser/OpParser.java
@@ -28,10 +28,11 @@ package jdk.incubator.code.parser;
 import java.io.IOException;
 import java.io.InputStream;
 import jdk.incubator.code.*;
-import jdk.incubator.code.ExternalizableTypeElement.*;
+import jdk.incubator.code.dialect.ExternalizableTypeElement.*;
 import jdk.incubator.code.dialect.DialectFactory;
 import jdk.incubator.code.dialect.ExternalizableOp;
 import jdk.incubator.code.dialect.OpFactory;
+import jdk.incubator.code.dialect.TypeElementFactory;
 import jdk.incubator.code.dialect.core.CoreOp;
 import jdk.incubator.code.dialect.core.CoreType;
 import jdk.incubator.code.dialect.java.JavaOp;
