+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/writer/OpBuilder.java
@@ -29,6 +29,7 @@ import jdk.incubator.code.*;
 import jdk.incubator.code.dialect.DialectFactory;
 import jdk.incubator.code.dialect.OpFactory;
 import jdk.incubator.code.dialect.ExternalizableOp;
+import jdk.incubator.code.dialect.TypeElementFactory;
 import jdk.incubator.code.dialect.core.CoreType;
 import jdk.incubator.code.dialect.core.FunctionType;
 import jdk.incubator.code.dialect.java.*;
